Refactor show images to new package
