+++ base/src/main/java/app/tivi/inject/Annotations.kt
@@ -31,6 +31,11 @@ annotation class Tmdb
 @Retention(AnnotationRetention.RUNTIME)
 @Qualifier
 @MustBeDocumented
+annotation class ForStore
+
+@Retention(AnnotationRetention.RUNTIME)
+@Qualifier
+@MustBeDocumented
 annotation class MediumDate
 
 @Retention(AnnotationRetention.RUNTIME)
