+++ data/src/main/java/app/tivi/data/repositories/showimages/TmdbShowImagesDataSource.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Google LLC
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package app.tivi.data.repositories.shows
+package app.tivi.data.repositories.showimages
 
 import app.tivi.data.entities.ErrorResult
 import app.tivi.data.entities.Result
