+++ domain/src/main/java/app/tivi/domain/interactors/UpdateFollowedShows.kt
@@ -23,7 +23,7 @@ import app.tivi.data.fetchCollection
 import app.tivi.data.instantInPast
 import app.tivi.data.repositories.episodes.SeasonsEpisodesRepository
 import app.tivi.data.repositories.followedshows.FollowedShowsRepository
-import app.tivi.data.repositories.shows.ShowImagesStore
+import app.tivi.data.repositories.showimages.ShowImagesStore
 import app.tivi.data.repositories.shows.ShowStore
 import app.tivi.data.repositories.watchedshows.WatchedShowsRepository
 import app.tivi.domain.Interactor
