+++ domain/src/main/java/app/tivi/domain/interactors/UpdateShowImages.kt
@@ -17,8 +17,8 @@
 package app.tivi.domain.interactors
 
 import app.tivi.data.fetchCollection
-import app.tivi.data.repositories.shows.ShowImagesLastRequestStore
-import app.tivi.data.repositories.shows.ShowImagesStore
+import app.tivi.data.repositories.showimages.ShowImagesLastRequestStore
+import app.tivi.data.repositories.showimages.ShowImagesStore
 import app.tivi.domain.Interactor
 import app.tivi.inject.ProcessLifetime
 import app.tivi.util.AppCoroutineDispatchers
